refactor: update @handbook/react
