--- public/frame.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#222232" />
-    <link rel="shortcut icon" href="favicon.ico" />
-    <link rel="manifest" href="manifest.json" />
-    <title>API</title>
-    <script src="cornerstone.js"></script>
-    <script src="cornerstoneWADOImageLoader.js"></script>
-    <script src="cornerstoneWebImageLoader.js"></script>
-    <script src="dicomParser.js"></script>
-    <link rel="stylesheet" href="fonts/proximanova/proximanova.css" />
-  </head>
-
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="app">Fuck!!!</div>
-  </body>
-</html>
