+++ src/_packages/@handbook/react/types.ts
@@ -1,11 +1,22 @@
-import { ComponentType, ReactNode } from 'react';
+import { ComponentType, DetailedHTMLProps, IframeHTMLAttributes } from 'react';
 
 export type PageContent = () => Promise<{ default: ComponentType }>;
 
 export type HandbookTreeNode = { [name: string]: HandbookTreeNode | PageContent };
 
 export interface HandbookConfig {
+  github?: {
+    repo: string;
+    branch?: string;
+  };
+
+  vscode?: boolean;
+
   index: HandbookTreeNode;
-  sampleTemplate?: string;
-  sampleWrapper?: ComponentType<{ children: ReactNode }>;
+
+  template?: string;
+
+  components?: {
+    IFrame?: ComponentType<DetailedHTMLProps<IframeHTMLAttributes<HTMLIFrameElement>, HTMLIFrameElement>>;
+  };
 }
