+++ src/_packages/@lunit/insight-viewer/__stories__/Annotation.Contour.mdx
@@ -2,6 +2,6 @@ import { Sample } from '@handbook/react';
 
 # `<ContourDrawer>`
 
-<Sample path="./Annotation.Contour.story" height={800}/>
+<Sample path="./Annotation.Contour.story" height={750}/>
 
 어쩌구 저쩌구...
