+++ src/_packages/@lunit/insight-viewer/__stories__/InsightViewer.mdx
@@ -2,6 +2,6 @@ import { Sample } from '@handbook/react';
 
 # `<InsightViewer>`
 
-<Sample path="./InsightViewer.story" height={800}/>
+<Sample path="./InsightViewer.story" height={750}/>
 
 어쩌구 저쩌구...
